# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Api documentation",
        "description": "Api documentation for to-do-app"
    },
    "host": "localhost:4000",
    "basePath": "",
    "tags": [
        {
            "name": "Auth",
            "description": "Auth Router"
        },
        {
            "name": "Todos",
            "description": "Todos"
        }
    ],
    "paths": {
        "/todo/getTodos": {
            "get": {
                "tags": [
                    "Todos"
                ],
                "summary": "Get all todos",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/todo"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/todo/getTodosByStatus": {
            "post": {
                "tags": [
                    "Todos"
                ],
                "summary": "Get all todos filtered by the status",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": "true",
                        "description": "Status object",
                        "schema": {
                            "$ref": "#/definitions/status"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/todo"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/todo/getTodoById/{todoId}": {
            "get": {
                "tags": [
                    "Todos"
                ],
                "summary": "Get todo by id",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "todoId",
                        "required": true,
                        "description": "todo id",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/todo"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/todo/deleteTodo/{todoId}": {
            "delete": {
                "tags": [
                    "Todos"
                ],
                "summary": "Delete todo by id",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "todoId",
                        "required": true,
                        "description": "todo object example",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/todo/updateTodoStatus/{todoId}": {
            "post": {
                "tags": [
                    "Todos"
                ],
                "summary": "Update the status of todo",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "todoId",
                        "required": true,
                        "description": "todo object example",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": "true",
                        "description": "Status object",
                        "schema": {
                            "$ref": "#/definitions/status"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/todo/addTodo": {
            "post": {
                "tags": [
                    "Todos"
                ],
                "summary": "Add new todo",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "description": "todo object example",
                        "schema": {
                            "$ref": "#/definitions/todo2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/todo"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login into the system",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "description": "Login object example",
                        "schema": {
                            "$ref": "#/definitions/user"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        }
    },
    "definitions": {
        "todo": {
            "properties": {
                "_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "planned",
                        "in progress",
                        "done",
                        "blocked"
                    ]
                },
                "responsable": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string",
                    "format": "date"
                },
                "finnishedDate": {
                    "type": "string",
                    "format": "date"
                }
            }
        },
        "todo2": {
            "properties": {
                "title": {
                    "type": "string"
                },
                "responsable": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string",
                    "format": "date"
                }
            }
        },
        "user": {
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "id": {
            "properties": {
                "id": {
                    "type": "number"
                }
            }
        },
        "msg": {
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "status": {
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "planned",
                        "in progress",
                        "done",
                        "blocked"
                    ]
                }
            }
        }
    }
}